requirement.txt
google-auth==2.18.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.83.0
yt-dlp==2024.10.0
schedule==1.1.0
